+++ trilby-cli/CHANGELOG.md
@@ -1,11 +1,13 @@
 # Revision history for trilby-cli
 
-## 24.5.0 -- 2024-05-01
+## 24.7.0 -- 2024-07-24
 
 * Print version when running verbosely
 * Use correct PATH when running in Nix sandbox
 * Update flakes recursively
 * Add Infect command
+* Switch from TH to callstack-based logging
+* Remove library, build only executable
 
 ## 24.4.0 -- 2024-04-04